Enable compatibility with Edge Runtime/browser (#21)
There are three items in our SDK (identified thus far) that do not work
on browser/Edge: Websockets, createVoice + fs import, and handling binary
data.

We have been using isomorphic-ws to enable websockets in both
browser and node environments. However, it appears that this library
will not work with the Edge Runtime. createVoice will not work
on browser (and presumably Edge) since we cannot directly access the
filesystem. Browser/Edge also uses Blobs, which we do not currently handle.
This causes problems with streaming, since we expect an array buffer.

This commit adds a runtime check function and returns the right
WebSocket object depending on the runtime. It also makes the fs import
conditional on the runtime, and prevent running createVoice outside of node
environments. Lastly, it adds logic to convert Blobs to buffers.